# docs/swagger.yaml
basePath: /
definitions:
  apistruct.ApiResponse:
    properties:
      data: {}
      errCode:
        example: 0
        type: integer
      errDlt:
        example: ""
        type: string
      errMsg:
        example: success
        type: string
    type: object
...
  apistruct.GetDesignateUsersReq:
    properties:
      userIDs:
        example:
        - user123
        - user456
        items:
          type: string
        type: array
    required:
    - userIDs
    type: object
  apistruct.GetDesignateUsersResp:
    properties:
      users:
        items:
          $ref: '#/definitions/apistruct.PublicUserInfo'
        type: array
    type: object
...
  apistruct.Int32Value:
    properties:
      value:
        type: integer
    type: object
  apistruct.Int64Value:
    properties:
      value:
        type: integer
    type: object
...
  apistruct.PublicUserInfo:
    properties:
      ex:
        example: '{}'
        type: string
      faceURL:
        example: https://example.com/avatar.jpg
        type: string
      gender:
        example: 1
        type: integer
      nickname:
        example: 张三
        type: string
      userID:
        example: user123
        type: string
    type: object
...
  apistruct.StringValue:
    properties:
      value:
        type: string
    type: object
  apistruct.UpdateUserInfoExReq:
    properties:
      areaCode:
        $ref: '#/definitions/apistruct.StringValue'
      birth:
        $ref: '#/definitions/apistruct.Int64Value'
      email:
        $ref: '#/definitions/apistruct.StringValue'
      ex:
        $ref: '#/definitions/apistruct.StringValue'
      faceURL:
        $ref: '#/definitions/apistruct.StringValue'
      gender:
        $ref: '#/definitions/apistruct.Int32Value'
      nickname:
        $ref: '#/definitions/apistruct.StringValue'
      phoneNumber:
        $ref: '#/definitions/apistruct.StringValue'
      userID:
        example: user123
        type: string
    required:
    - userID
    type: object
  apistruct.UserInfo:
    properties:
      areaCode:
        example: "+86"
        type: string
      birth:
        example: 946656000000
        type: integer
      email:
        example: user@example.com
        type: string
      ex:
        example: '{}'
        type: string
      faceURL:
        example: https://example.com/avatar.jpg
        type: string
      gender:
        example: 1
        type: integer
      nickname:
        example: 张三
        type: string
      phoneNumber:
        example: "13812345678"
        type: string
      userID:
        example: user123
        type: string
    required:
    - nickname
    - userID
    type: object
  apistruct.UserRegisterReq:
    properties:
      secret:
        example: openIM123
        type: string
      users:
        items:
          $ref: '#/definitions/apistruct.UserInfo'
        type: array
    required:
    - secret
    - users
    type: object
  apistruct.UserRegisterResp:
    properties:
      userID:
        example: user123
        type: string
    type: object
host: localhost:10002
info:
  contact:
    email: contact@openim.io
    name: OpenIM Support
    url: https://github.com/openimsdk/open-im-server
  description: |-
    OpenIM Server 提供即时通讯服务的 REST API 接口文档

    ## 认证方式
    大部分接口需要在 Header 中携带 token 进行认证：
    - Header名称: token
    - 获取方式: 调用 /auth/get_user_token 接口

    ## 接口规范
    - 所有接口均使用 POST 方法
    - Content-Type: application/json
    - 响应格式: JSON
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenIM Server API
  version: "3.8"
paths:
...
  /user/get_users_info:
    post:
      consumes:
      - application/json
      description: 批量获取指定用户的公开信息
      parameters:
      - description: 用户ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.GetDesignateUsersReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应，返回用户信息列表
          schema:
            allOf:
            - $ref: '#/definitions/apistruct.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/apistruct.GetDesignateUsersResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 获取用户公开信息
      tags:
      - 用户管理
  /user/update_user_info_ex:
    post:
      consumes:
      - application/json
      description: 更新用户的详细信息，支持部分更新
      parameters:
      - description: 用户信息更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.UpdateUserInfoExReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 更新用户信息
      tags:
      - 用户管理
  /user/user_register:
    post:
      consumes:
      - application/json
      description: 注册新用户账号
      parameters:
      - description: 用户注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.UserRegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/apistruct.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/apistruct.UserRegisterResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "409":
          description: 用户已存在
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 用户注册
      tags:
      - 用户管理
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"